core/requirements_pod/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
alembic>=1.13.0
PyMySQL>=1.1.0
cloud-sql-python-connector[pymysql]>=1.9.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
httpx>=0.27.0
anthropic>=0.28.0
python-multipart>=0.0.9
python-docx>=1.1.0
PyMuPDF>=1.24.0
google-cloud-storage>=2.16.0
python-dotenv>=1.0.0
claude-agent-sdk
anyio>=4.0.0
pytest>=8.2.0
pytest-asyncio>=0.23.0
respx>=0.21.0
